requirements.txt
# GVP Watch Backend - Python Dependencies
# ==========================================
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1

# HTTP Client
httpx==0.25.2

# Job Scheduling
apscheduler==3.10.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-httpx==0.27.0

# Logging & Monitoring
python-json-logger==2.0.7

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
